+++ test/e2e/operation/pipeline/src/test/java/org/apache/shardingsphere/test/e2e/operation/pipeline/dao/orderitem/sqlbuilder/MySQLIntPkOrderItemSQLBuilder.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.test.e2e.operation.pipeline.dao.orderitem.sqlbuilder;
 
-public final class MySQLOrderItemSQLBuilder implements OrderItemSQLBuilder {
+public final class MySQLIntPkOrderItemSQLBuilder implements IntPkOrderItemSQLBuilder {
     
     @Override
     public String buildCreateTableSQL() {
